cifar_10_image_classification/requirements.txt
fastapi
uvicorn
torch
torchvision
onnx
pillow
pyngrok
nest_asyncio
python-multipart
